+++ app/api/temp_messages/messages/api.rb
@@ -1,0 +1,9 @@
+module TempMessages
+  module Messages
+    class API < Grape::API
+      prefix '/messages'
+      mount Post
+      mount Get
+    end
+  end
+end
