+++ config/routes.rb
@@ -2,7 +2,7 @@ Rails.application.routes.draw do
   # Define your application routes per the DSL in https://guides.rubyonrails.org/routing.html
 
   # Defines the root path route ("/")
-  mount Messages::API => '/'
+  mount TempMessages::API => '/api'
   root 'messages#new'
 
   resources :messages, only: %i[show new create]
