+++ spec/requests/temp_messages/messages/get_spec.rb
@@ -1,19 +1,19 @@
 require 'rails_helper'
 
 RSpec.describe 'Messages API', type: :request do
-  describe 'GET /api/messages/:id' do
+  describe 'GET /api/temp-messages/messages/:id' do
     let(:messages) { create_list(:message, 5) }
 
     context 'when the record exists' do
       let(:sample_message) { messages.sample }
 
       it 'returns ok status' do
-        get "/api/messages/#{sample_message.id}"
+        get "/api/temp-messages/messages/#{sample_message.id}"
         expect(response).to have_http_status(:ok)
       end
 
       it "returns the message's content" do
-        get "/api/messages/#{sample_message.id}"
+        get "/api/temp-messages/messages/#{sample_message.id}"
         expect(JSON.parse(response.body)['message']['content']).not_to be_empty
       end
     end
@@ -22,12 +22,12 @@ RSpec.describe 'Messages API', type: :request do
       let(:message_id) { 1000 }
 
       it 'returns status code 404' do
-        get "/api/messages/#{message_id}"
+        get "/api/temp-messages/messages/#{message_id}"
         expect(response).to have_http_status(:not_found)
       end
 
       it 'returns an error message' do
-        get "/api/messages/#{message_id}"
+        get "/api/temp-messages/messages/#{message_id}"
         expect(JSON.parse(response.body)['error']).not_to be_empty
       end
     end
@@ -40,12 +40,12 @@ RSpec.describe 'Messages API', type: :request do
       end
 
       it 'returns status code 401' do
-        get "/api/messages/#{expired_message.id}"
+        get "/api/temp-messages/messages/#{expired_message.id}"
         expect(response).to have_http_status(:unauthorized)
       end
 
       it 'returns an error message' do
-        get "/api/messages/#{expired_message.id}"
+        get "/api/temp-messages/messages/#{expired_message.id}"
         expect(JSON.parse(response.body)['error']).not_to be_empty
       end
     end
@@ -53,17 +53,19 @@ RSpec.describe 'Messages API', type: :request do
     context 'when the record has an invalid expiration type' do
       let(:invalid_message) { messages.last }
 
+      # rubocop:disable Rails/SkipsModelValidations
       before do
         invalid_message.update_column(:expiration_type, 'invalid')
       end
+      # rubocop:enable Rails/SkipsModelValidations
 
       it 'returns status code 422' do
-        get "/api/messages/#{invalid_message.id}"
+        get "/api/temp-messages/messages/#{invalid_message.id}"
         expect(response).to have_http_status(:unprocessable_entity)
       end
 
       it 'returns an error message' do
-        get "/api/messages/#{invalid_message.id}"
+        get "/api/temp-messages/messages/#{invalid_message.id}"
         expect(JSON.parse(response.body)['error']).not_to be_empty
       end
     end
